requirements.txt
# OP Trading Platform - Complete Requirements
# ===========================================
# Version: 3.1.2 - Production Dependencies
# Date: 2025-08-25 10:50 PM IST

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web Framework Support
starlette==0.27.0
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Database and Storage
influxdb-client==1.39.0
redis==5.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# Data Processing and Analytics
pandas==2.1.4
numpy==2.1.0
scipy==1.11.4
scikit-learn==1.3.2
pytz==2023.3

# Trading and Market Data
kiteconnect==4.1.0
yfinance==0.2.18
ta==0.10.2
python-dateutil==2.8.2

# Monitoring and Metrics
prometheus-client==0.19.0
psutil==5.9.6
structlog==23.2.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2
configparser==6.0.0

# Authentication and Security
cryptography==41.0.8
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.92.1
factory-boy==3.3.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
sphinx==7.2.6

# Data Validation and Serialization
marshmallow==3.20.1
cerberus==1.3.5
jsonschema==4.20.0

# Async and Concurrency
asyncio==3.4.3
aiofiles==23.2.1
aiocache==0.12.2

# Networking and HTTP
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17

# Utilities
click==8.1.7
rich==13.7.0
python-multipart==0.0.6
email-validator==2.1.0
phonenumbers==8.13.25

# Data Export and Import
openpyxl==3.1.2
xlsxwriter==3.1.9
fpdf2==2.7.6

# Background Tasks and Queuing
celery==5.3.4
kombu==5.3.4

# Caching
python-memcached==1.59
diskcache==5.6.3

# Logging and Debugging
loguru==0.7.2
sentry-sdk==1.38.0

# Performance and Optimization
cython==3.0.6
orjson==3.9.10
msgpack==1.0.7

# Cloud and Infrastructure
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# Development Dependencies (Optional)
jupyterlab==4.0.8
notebook==7.0.6
ipython==8.18.1

# System Integration
watchdog==3.0.0
schedule==1.2.0
apscheduler==3.10.4

# Data Visualization (Backend)
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# API Documentation
redoc==2.1.0
swagger-ui-bundle==0.0.9

# Error Handling and Recovery
tenacity==8.2.3
backoff==2.2.1

# Locale and Internationalization
babel==2.13.1
pyicu==2.11

# Hardware and System Monitoring
nvidia-ml-py==12.535.133
gpustat==1.1.1

# Development Quality Tools
bandit==1.7.5
safety==2.3.4
pip-audit==2.6.1

# Mock and Testing Data
faker==20.1.0
responses==0.24.1
requests-mock==1.11.0

# Advanced Analytics (Optional)
ta-lib==0.4.28
backtrader==1.9.78.123
zipline-reloaded==3.0.3

# Notification Services
twilio==8.11.0
slack-sdk==3.26.1
sendgrid==6.11.0

# File Processing
python-magic==0.4.27
pillow==10.1.0
pdfplumber==0.10.3

# Encryption and Hashing
hashlib2==1.0.2
pyotp==2.9.0
qrcode==7.4.2

# API Rate Limiting
slowapi==0.1.9
limits==3.6.0

# Background Processing
dramatiq==1.15.0
huey==2.5.0

# Development Containers
docker==6.1.3
docker-compose==1.29.2

# Time Series and Financial Analysis
arch==6.2.0
statsmodels==0.14.0
pmdarima==2.0.4

# Machine Learning (Optional)
tensorflow==2.15.0
torch==2.1.2
xgboost==2.0.2

# Natural Language Processing
textblob==0.17.1
nltk==3.8.1
spacy==3.7.2